Validate company admin

+++ app/Livewire/Employee/Index.php
@@ -2,6 +2,7 @@
 
 namespace App\Livewire\Employee;
 
+use App\Enums\RoleEnum;
 use App\Livewire\Traits\Roles;
 use App\Livewire\Traits\Table;
 use App\Models\Department;
@@ -90,6 +91,12 @@ class Index extends Component
     public function destroyEmployee(): void
     {
         $employee = User::find($this->selected_employee_id);
+        $system_owner = $employee->hasRole(RoleEnum::COMPANY_ADMIN->value);
+        if ($system_owner) {
+            Flux::modal('confirm-destroy-employee-modal')->close();
+            $this->dispatch('toast', message: __('No se puede eliminar al administrador de la empresa.'), type: 'error');
+            return;
+        }
         $employee->delete();
 
         $this->table_items = array_filter($this->table_items, fn($item) => $item['id'] !== $this->selected_employee_id);
